sidebar: split out window.c
Split the sidebar window code into a new source file.
This encapsulates `recalc()` and `repaint()` functions.
